Replace loop-lit source code with the .ll files
I don't think there is any point to keep the .c and .i files in our repo since we are
not planning to change them and then we are just forced to merge with the
sv-benchmarks repo every so often. Instead we should just keep the .ll files
that we want to verify and can use as the basis of tests.