Fix term generation bug due to referencing array names and index values at the wrong nodes
Because the array and index nodes come from the ArrayElement matcher which gets them
from the code properties, they are not in the trace. Accessing the stores chain from them
gets the wrong values. Added versions of nterm and vterm that also take a node at which
to access the change to look up any names.