Refactor Main and CMain to use configurable front-ends instead of IR providers
What was formerly an IR provider is now a Frontend and there is only one Main now. CMain is gone. The main program now takes a -f option that can be used to select the front-end to use (default is C). Main uses the selected front-end to process the input file to obtain an IR representation which is then processed for verification as before.

The old -i option for selecting the IR provider is gone since it is subsumed but the choice of front-end. It's up to the front-end to return an IR representation appropriate to that front-end (always LLVM IR for now).

There are three front-end implementations: IML, C and LLVM. The IML one compiles the IML code directly to LLVM IR. The C one runs clang externally to get an LLVM IR file, then uses the LLVM front-end to make the IR representation. The LLVM front-end just parses an LLVM IR file. The front-ends now all log similar information.