Make the solver command-line option actually affect solver choice
Previously we always used Z3 regardless of how the -e option was set. Now the option
selects a solver mode which is then used by the trace refinement algorithm to instantiate
the solver. This approach is being used since in the near future we will want to also
look at the integer mode so select the theory. It's best to do that in the verifier rather
than as part of option processing.