Refactor itemTerm so that we get a termination if there is a global var whose initialisation we don’t handle
Previously we would just silently generate true for ones we didn’t handle which lead to the possibility that a semantics-affecting part of the program would be ignored without us noticing. Now we get a crash so we have to explicitly deal with it, either to put in a case to ignore it or to generate an appropriate term.