Add support for dummy correctness witnesses
It turns out that if we don’t generate a correctness witness at all then the validation fails and
benchexec doesn’t seem to even give us credit for getting the TRUE verification result. These
changes cause Skink to output witness for TRUE results. After some experimentation I’ve
settled on a dummy witness that simply provides a true invariant. CPAchecker seems to be
happy with this as a valid correctness witness. It doesn’t validate it, but at least it doesn’t
error out now *and* we get the credit for the TRUE result.