Upgrade bench script to include witness validation and HTML table preparation
See Vagrant.md for details on how to use this. You’ll need to run in a VM that is
provisioned by the new Vagrantfile.

Basically bench now runs benchexec twice: once to run Skink to verify the chosen
benchmarks and once to run CPAchecker. It also merges the results from the two
runs and generates a HTML table summary of the merged results. The new
skink.xml is almost the same as the one in the sv-comp repo.