Make the function namer do the right thing for the RHS of phi instructions and test it
Basically the RHSes have to be interpreted as accessing values on the incoming edges. So if
one phi insn refers to a variable set by an earlier one, it should get the previous value of the
variable not the value set by the earlier phi insn. This change makes the namer use the
incoming index map of the enclosing block for such cases. Also added a test case.