Fix a bug where the term builder disagreed with the control flow graph about switch choices
The convention is now consistently used that if the switch has n cases then choices 0..n-1 are the choices that corresponds to the cases of the switch and choice n corresponds to the default choice.

Previously the term builder was using choice zero for the default case and 1..n for the cases but this disagreed with the control flow graph.

One reason to prefer the current scheme is that it means if we process the choices in order (which seems likely) then we will process “interesting” choices first and the default (uninteresting) case last. Whether this makes a real difference in benchmarks is unclear.