First attempt at a fallback strategy
Since we are no longer allowed to have category-specific arguments, we need to move to a fallback
approach if we can't tell from the benchmark code what options to use. skink,sh now has no
arguments except the program to verify. skink.xml now reflects that too.

The verifier now runs a series of configurations made up from the command-line arguments augmented
with the configuration arguments (see Verifier.scala).

Some things here are less than ideal, particularly: the verifier breaks the IR/LLVM IR barrier because
it needs to make new IRFunctions for each configuration. In the current design the command-line
config is bound up in the IRFunction so that function won't do.