Remove fesetround and fegetround support for now
For now we just say UNKNOWN if a program has a call to one of these
functions. This is desirable since Clang does not optimise in full awareness
of changes to the rounding mode.
Without this change Skink gets a number of benchmarks wrong.

We're still exploring whether some of the benchmarks have undefined behaviour
and hence should be removed.