Make term build more precise about what getelementptr insns it accepts
We can really only deal with simple single-dimenion array accesses. The
previous version causes problems when a pointer is computed from
another array access. Skink loses track of the fact that the two addresses
reference into the same array.  E.g., for the following program we get
the wrong answer:

array-multidimensional/init-non-constant-2-n-u_true-unreach-call.i

With this change we give UNKNOWN for this program, and seem to
be given the right results for Arrays programs that we got right
before the change.