Use QF_AUFLIA instead of AUFNIRA
This removes support for Real arithmetic on the basis that this version
of Skink is not good at that anyway and makes some wrong decisions
when using Real as opposed to FPBV. Now it will exit, and we can use
FPBV-capable Skink to pick up those benchmarks.

One possible regression is that this shifts to linear integer arithmetic
so maybe there are cases where we were benefiting from
non-linear arithmetic before which we can't now. But it's not clear if there is a QF_AUFNIA.