Improve capture of log
It seems that if the first Skink run was terminated for some reason
then the log file from benchexec would be empty. We now tee the
output into the results file so it is captured as we go.

Also, removed duplication of the code to copy the witness file.

Added more safeguard to delete results file when we are done, but it
doesn't seem to always work...