Reverse previous change
Of course we need clang, solvers, etc for CITests etc, so we need the
more advanced docker image.

Interesting to note is that hte CITests fail with "CITests.(It is not
a test it is a sbt.testing.SuiteSelector)" error using this image too.
So it's probably not an sbt setup problem.