Fixes for auxiliary SMT variable indexing
Auxiliary variables are used to encode memory operations. E.g., a $o
variable holds the current offset of an address into its memory chunk.
Similarly for $b for the bits of a value and $p, $e and $s for the
parts of a floating-point value.

These changes fix problems whereby these variables were not being correctly
indexed for different stores to the same LLVM variable. We now use the
same index as the related LLVM variable with the suffix appended. We
are also now more careful to reuse these names. Finally, we now more correctly
propagate offsets between
stores.