Fix handling of non-supported solver x mode
If some solvers don't support both bit and math mode we would get a
crash if we requested that solver in that mode, even if another solver
did support that mode. Now we take the requested solvers and
filter them to just those that support the mode. It's only an error if
none are left.