Add first version of solver guessing
If the user specifies which solvers to use, we use them (or at least the
ones that support the number mode). If the user doesn't specify, we
first try to guess which solver to use. If a guess applies, we use that
solver, otherwise we use the default collection.

Currently guessing is done on a trace-by-trace basis and we look at the terms
that encode the trace. The only one that's implemented currently is if
the trace includes only Bool and FPBV sorts then we use mathsat in FPBV
mode. This is because that mode can't be trusted when normal BVs are
around as well.