Add an optimisation for assume calls
Clang turns assume calls into a computation of an i1 for the condition
then a zero extension of that to i32 then a call to the assume function.
This means that we use a BV to encode the 32 bit version of the
condition, even though a Bool would do.

Here we add an extra pass to makeVerifiable to turn the clang code
into a single instruction that just calls the assume function with the
i1. Thus, the term builder will use a Bool for that value.

The immediate application is to programs that use FP operations
and we are in BV mode. To be able to use the fast mathsat-fp
solver, we don't want any BVs to actually be used. This transformation
means that the assume ones at least are turned into Bool. All in all,
this enables us to solve some FP benchmarks quickly using mathsat-fp.