Add logback configuration to main classpath
Previously we were relying on the default logging config when we run the
Skink application (e.g. in the competition). This logged a huge amount
of information, much of which wasn't useful (e.g., reindexing symbols
from scalasmt).

Now we can configure it properly. Currently it is set to log only the most useful
stuff and all to STDOUT.

Note that the sbt "run" task is setup in build.sbt to use the testing
logback-test.xml, so it is configured separately.